@@ -1,52 +1,72 @@
 {
     "name": "monsieurbiz/sylius-sales-reports-plugin",
     "type": "sylius-plugin",
-    "keywords": ["sylius", "sylius-plugin"],
+    "keywords": ["sylius", "sylius-plugin", "monsieurbiz"],
     "description": "A simple plugin to have sales reports in Sylius",
     "license": "MIT",
     "require": {
-        "php": "~7.2",
-        "sylius/sylius": "~1.3"
+        "php": "~7.4|~8.0",
+        "sylius/sylius": ">=1.8 <1.11"
     },
     "require-dev": {
-        "behat/behat": "^3.4",
-        "behat/mink": "^1.7@dev",
-        "behat/mink-browserkit-driver": "^1.3",
-        "behat/mink-extension": "^2.2",
-        "behat/mink-selenium2-driver": "^1.3",
+        "behat/behat": "^3.6.1",
+        "behat/mink-selenium2-driver": "^1.4",
+        "dmore/behat-chrome-extension": "^1.3",
+        "dmore/chrome-mink-driver": "^2.7",
+        "doctrine/data-fixtures": "^1.4",
+        "ergebnis/composer-normalize": "^2.5",
+        "friends-of-behat/mink": "^1.8",
+        "friends-of-behat/mink-browserkit-driver": "^1.4",
+        "friends-of-behat/mink-extension": "^2.4",
         "friends-of-behat/page-object-extension": "^0.3",
-        "friends-of-behat/suite-settings-extension": "^1.0",
-        "friends-of-behat/symfony-extension": "^2.0",
-        "friends-of-behat/variadic-extension": "^1.1",
-        "lakion/mink-debug-extension": "^1.2.3",
-        "phpspec/phpspec": "^5.0",
-        "phpstan/phpstan-doctrine": "^0.11",
-        "phpstan/phpstan-shim": "^0.11",
-        "phpstan/phpstan-webmozart-assert": "^0.11",
-        "phpunit/phpunit": "^8.0",
-        "sensiolabs/security-checker": "^5.0",
-        "sylius-labs/coding-standard": "^3.0",
-        "symfony/browser-kit": "^3.4|^4.3",
-        "symfony/debug-bundle": "^3.4|^4.3",
-        "symfony/dotenv": "^4.3",
-        "symfony/intl": "^3.4|^4.3",
-        "symfony/web-profiler-bundle": "^3.4|^4.3",
-        "symfony/web-server-bundle": "^3.4|^4.3"
+        "friends-of-behat/symfony-extension": "^2.1",
+        "friends-of-behat/variadic-extension": "^1.3",
+        "hwi/oauth-bundle": "^1.1",
+        "lchrusciel/api-test-case": "^5.0",
+        "matthiasnoback/symfony-config-test": "^4.2",
+        "matthiasnoback/symfony-dependency-injection-test": "^4.1",
+        "mikey179/vfsstream": "^1.6",
+        "mockery/mockery": "^1.4",
+        "pamil/prophecy-common": "^0.1",
+        "phpspec/phpspec": "^6.1",
+        "phpstan/phpstan": "^0.12.57",
+        "phpstan/phpstan-doctrine": "^0.12.19",
+        "phpstan/phpstan-webmozart-assert": "^0.12.7",
+        "phpunit/phpunit": "^8.5",
+        "psalm/plugin-mockery": "^0.3",
+        "psr/event-dispatcher": "^1.0",
+        "sylius-labs/coding-standard": "^3.1",
+        "symfony/browser-kit": "^4.4",
+        "symfony/debug-bundle": "^4.4",
+        "symfony/dotenv": "^4.4",
+        "symfony/flex": "^1.7",
+        "symfony/web-profiler-bundle": "^4.4",
+        "phpmd/phpmd": "@stable"
+    },
+    "conflict": {
+        "doctrine/dbal": "^3"
     },
     "prefer-stable": true,
     "autoload": {
         "psr-4": {
-            "MonsieurBiz\\SyliusSalesReportsPlugin\\": "src/",
-            "Tests\\MonsieurBiz\\SyliusSalesReportsPlugin\\": "tests/"
+            "MonsieurBiz\\SyliusSalesReportsPlugin\\": "src/"
         }
     },
-    "autoload-dev": {
-        "psr-4": {
-            "Tests\\MonsieurBiz\\SyliusSalesReportsPlugin\\": "tests/"
+    "scripts": {
+        "auto-scripts": {
+            "cache:clear": "symfony-cmd",
+            "assets:install %PUBLIC_DIR%": "symfony-cmd"
         },
-        "classmap": ["tests/Application/Kernel.php"]
+        "phpcs": "php-cs-fixer fix --using-cache=false",
+        "phpstan": "phpstan analyse -c phpstan.neon src/",
+        "phpmd": "phpmd --exclude Migrations/* src/ ansi phpmd.xml",
+        "phpunit": "phpunit",
+        "phpspec": "phpspec run"
     },
     "extra": {
+        "symfony": {
+            "require": "^4.4"
+        },
         "branch-alias": {
             "dev-master": "1.0-dev"
         }
