@@ -1,5 +1,14 @@
 <?php
 
+/*
+ * This file is part of Monsieur Biz' Sales Reports plugin for Sylius.
+ *
+ * (c) Monsieur Biz <sylius@monsieurbiz.com>
+ *
+ * For the full copyright and license information, please view the LICENSE.txt
+ * file that was distributed with this source code.
+ */
+
 declare(strict_types=1);
 
 namespace MonsieurBiz\SyliusSalesReportsPlugin;
@@ -15,9 +24,9 @@ final class MonsieurBizSyliusSalesReportsPlugin extends Bundle
     /**
      * Returns the plugin's container extension.
      *
-     * @return ExtensionInterface|null The container extension
-     *
      * @throws \LogicException
+     *
+     * @return ExtensionInterface|null The container extension
      */
     public function getContainerExtension(): ?ExtensionInterface
     {
@@ -26,13 +35,14 @@ final class MonsieurBizSyliusSalesReportsPlugin extends Bundle
 
             if (null !== $extension) {
                 if (!$extension instanceof ExtensionInterface) {
-                    throw new \LogicException(sprintf('Extension %s must implement %s.', get_class($extension), ExtensionInterface::class));
+                    throw new \LogicException(sprintf('Extension %s must implement %s.', \get_class($extension), ExtensionInterface::class));
                 }
                 $this->containerExtension = $extension;
             } else {
                 $this->containerExtension = false;
             }
         }
+
         return $this->containerExtension instanceof ExtensionInterface
             ? $this->containerExtension
             : null;
