@@ -1,5 +1,14 @@
 <?php
 
+/*
+ * This file is part of Monsieur Biz' Sales Reports plugin for Sylius.
+ *
+ * (c) Monsieur Biz <sylius@monsieurbiz.com>
+ *
+ * For the full copyright and license information, please view the LICENSE.txt
+ * file that was distributed with this source code.
+ */
+
 declare(strict_types=1);
 
 namespace MonsieurBiz\SyliusSalesReportsPlugin\DependencyInjection;
@@ -12,10 +21,9 @@ use Symfony\Component\DependencyInjection\Loader\YamlFileLoader;
 
 final class MonsieurBizSyliusSalesReportsExtension extends Extension
 {
+    public const EXTENSION_CONFIG_NAME = 'monsieurbiz_sylius_sales_reports';
 
-    CONST EXTENSION_CONFIG_NAME = 'monsieurbiz_sylius_sales_reports';
-
-    public function load(array $configs, ContainerBuilder $container)
+    public function load(array $configs, ContainerBuilder $container): void
     {
         $configuration = $this->getConfiguration([], $container);
         if ($configuration instanceof ConfigurationInterface) {
