@@ -1,7 +1,18 @@
 <?php
 
+/*
+ * This file is part of Monsieur Biz' Sales Reports plugin for Sylius.
+ *
+ * (c) Monsieur Biz <sylius@monsieurbiz.com>
+ *
+ * For the full copyright and license information, please view the LICENSE.txt
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
 namespace MonsieurBiz\SyliusSalesReportsPlugin\Exception;
 
-class InvalidDateException extends \Exception {
-
+class InvalidDateException extends \Exception
+{
 }
