@@ -1,5 +1,16 @@
 <?php
 
+/*
+ * This file is part of Monsieur Biz' Sales Reports plugin for Sylius.
+ *
+ * (c) Monsieur Biz <sylius@monsieurbiz.com>
+ *
+ * For the full copyright and license information, please view the LICENSE.txt
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
 namespace MonsieurBiz\SyliusSalesReportsPlugin\Menu;
 
 use Knp\Menu\ItemInterface;
@@ -8,9 +19,7 @@ use Sylius\Bundle\UiBundle\Menu\Event\MenuBuilderEvent;
 class AdminMenuListener
 {
     /**
-     * Add reports link in sales menu
-     *
-     * @param MenuBuilderEvent $event
+     * Add reports link in sales menu.
      */
     public function addAdminMenuItem(MenuBuilderEvent $event): void
     {
